# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Condominium Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies/{companyId}/invitations:
    get:
      operationId: getCompanyInvitations
      tags: [invitations]
      summary: List all pending invitations for a company
      parameters:
        - $ref: "#/components/parameters/companyId"
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, active, revoked, all]
          description: Filter by status (default pending)
      responses:
        "200":
          description: List of invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvitationDetail"
...
  /units/{unitId}/invite-owner:
    post:
      operationId: inviteOwner
      tags: [invitations]
      summary: Invite an owner to a unit (administrator only)
      parameters:
        - $ref: "#/components/parameters/unitId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerInvitationInput"
      responses:
        "201":
          description: Owner invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitMembership"
        "409":
          description: Active or pending owner already exists for this unit
  /units/{unitId}/invite-tenant:
    post:
      operationId: inviteTenant
      tags: [invitations]
      summary: Invite a tenant to a unit (owner only)
      parameters:
        - $ref: "#/components/parameters/unitId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInvitationInput"
      responses:
        "201":
          description: Tenant invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitMembership"
        "409":
          description: Active tenant or pending tenant invitation already exists

  # ── Unit Memberships ───────────────────────────────────────────────────────
  /unit-memberships/{membershipId}:
    delete:
      operationId: revokeUnitMembership
      tags: [memberships]
      summary: Revoke a unit membership (tenant revocation or invitation cancellation)
      parameters:
        - $ref: "#/components/parameters/membershipId"
      responses:
        "200":
          description: Membership revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitMembership"
  /unit-memberships/{membershipId}/resend:
    post:
      operationId: resendInvitation
      tags: [invitations]
      summary: Resend a pending invitation
      parameters:
        - $ref: "#/components/parameters/membershipId"
      responses:
        "200":
          description: Invitation resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitMembership"

  # ── Invitations ────────────────────────────────────────────────────────────
  /invitations/accept:
    post:
      operationId: acceptInvitation
      tags: [invitations]
      summary: Accept an invitation using a token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationAcceptInput"
      responses:
        "200":
          description: Invitation accepted, membership activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationAcceptResult"
        "400":
          description: Token invalid or expired
        "409":
          description: Invitation already accepted

  # ── Me (self-service) ──────────────────────────────────────────────────────
...
components:
  parameters:
    companyId:
      name: companyId
      in: path
      required: true
      schema:
        type: string
...
    unitId:
      name: unitId
      in: path
      required: true
      schema:
        type: string
    membershipId:
      name: membershipId
      in: path
      required: true
      schema:
        type: string

# ── Schemas ───────────────────────────────────────────────────────────────────
  schemas:
...
    Building:
      type: object
      required: [id, companyId, name, locality, status, createdAt]
      properties:
        id:
          type: string
        companyId:
          type: string
        name:
          type: string
        addressLine1:
          type: ["string", "null"]
        addressLine2:
          type: ["string", "null"]
        locality:
          type: string
        postcode:
          type: ["string", "null"]
        country:
          type: string
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Unit:
      type: object
      required: [id, companyId, buildingId, unitNumber, unitType, status, createdAt]
      properties:
        id:
          type: string
        companyId:
          type: string
        buildingId:
          type: string
        unitNumber:
          type: string
        unitType:
          type: string
          enum: [apartment, garage, commercial, other]
        floor:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, archived]
        activatedAt:
          type: ["string", "null"]
          format: date-time
        archivedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    OwnerInvitationInput:
      type: object
      required: [invitedName, invitedEmail]
      properties:
        invitedName:
          type: string
          minLength: 1
        invitedEmail:
          type: string
    TenantInvitationInput:
      type: object
      required: [invitedName, invitedEmail]
      properties:
        invitedName:
          type: string
          minLength: 1
        invitedEmail:
          type: string
    InvitationAcceptInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 1
    UnitMembership:
      type: object
      required: [id, unitId, role, status, invitedName, invitedEmail, createdAt]
      properties:
        id:
          type: string
        unitId:
          type: string
        userId:
          type: ["string", "null"]
        role:
          type: string
          enum: [owner, tenant]
        status:
          type: string
          enum: [pending, active, revoked]
        invitedName:
          type: string
        invitedEmail:
          type: string
        invitationExpiresAt:
          type: ["string", "null"]
          format: date-time
        activatedAt:
          type: ["string", "null"]
          format: date-time
        revokedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    InvitationDetail:
      allOf:
        - $ref: "#/components/schemas/UnitMembership"
        - type: object
          required: [unit, building]
          properties:
            unit:
              $ref: "#/components/schemas/Unit"
            building:
              $ref: "#/components/schemas/Building"
    InvitationAcceptResult:
      type: object
      required: [membership, role]
      properties:
        membership:
          $ref: "#/components/schemas/UnitMembership"
        role:
          type: string
          enum: [owner, tenant]
        redirectTo:
          type: string
          description: Suggested frontend redirect path after acceptance